LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-forms/appearance-base-basic-expected.txt

     button

FAIL Basic appearance stylesheet properties for input-text assert_equals: color should inherit expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Basic appearance stylesheet properties for input-search assert_equals: color should inherit expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Basic appearance stylesheet properties for input-number assert_equals: color should inherit expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Basic appearance stylesheet properties for input-button assert_equals: color should inherit expected "rgb(0, 128, 0)" but got "rgb(0, 136, 255)"
FAIL Basic appearance stylesheet properties for input-reset assert_equals: color should inherit expected "rgb(0, 128, 0)" but got "rgb(0, 136, 255)"
FAIL Basic appearance stylesheet properties for button assert_equals: color should inherit expected "rgb(0, 128, 0)" but got "rgb(0, 136, 255)"
PASS Basic appearance stylesheet properties for select
FAIL Basic appearance stylesheet properties for textarea assert_equals: color should inherit expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
PASS Basic appearance stylesheet properties for meter
PASS Basic appearance stylesheet properties for progress
